FQMA/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
Flask-SocketIO==5.5.1

langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-experimental==0.3.4
langchain-neo4j==0.5.0
langchain-openai==0.3.28

neo4j==5.28.1
psycopg2-binary==2.9.10
PyMySQL==1.1.1

rdflib==7.1.4
python-dotenv==1.1.1

numpy==2.3.1